requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.4.0
tqdm>=4.65.0
pillow>=9.0.0
wandb>=0.15.0
scikit-learn>=1.0.0
pathlib>=1.0.1
gdown